src/requirements/common.txt
alembic==1.9.3
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
dominate==2.7.0
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-Migrate==4.0.4
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
py4j==0.10.9.5
pyarrow==11.0.0
pyrsistent==0.19.3
pyspark==3.3.2
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
spark==0.2.1
SQLAlchemy==2.0.3
tomli==2.0.1
typing_extensions==4.5.0
visitor==0.1.3
Werkzeug==2.2.3
